Consolidate version-independent property tests
